refactor(#5) :: 사용자 정보 변경 모듈화
