+++ src/containers/ui/profileEdit/index.tsx
@@ -1,36 +1,18 @@
 import * as _ from "./style";
-import { useUserStore } from "@/store/user";
 import Image from "next/image";
 import { useRouter } from "next/navigation";
-import { useState } from "react";
 import Btn from "@/components/button";
+import { useChangeUserInfo } from "@/hooks/changeUserInfo";
 
 const ProfileEdit = () => {
-  const user = useUserStore((s) => s.user);
-  const setUser = useUserStore((s) => s.setUser);
-  const [name, setName] = useState(user.name);
-  const [mail, setMail] = useState(user.mail);
-  const userInput = [
-    {
-      name: "이름",
-      value: name,
-      setValue: setName,
-    },
-    {
-      name: "계정",
-      value: mail,
-      setValue: setMail,
-    },
-  ];
   const router = useRouter();
+  const { user, userInput, apply } = useChangeUserInfo();
+
   const handleSubmit = () => {
-    setUser({
-      img: user.img,
-      name: name,
-      mail: mail,
-    });
+    apply(); // 훅에서 스토어 업데이트
     router.push("/profile");
   };
+
   return (
     <_.Container>
       <_.BasicSet>
@@ -46,7 +28,7 @@ const ProfileEdit = () => {
             <_.EditItem key={i}>
               <_.EditText>{item.name}</_.EditText>
               <_.EditInput
-                type={"text"}
+                type="text"
                 value={item.value}
                 onChange={(e) => item.setValue(e.target.value)}
               />
@@ -58,4 +40,5 @@ const ProfileEdit = () => {
     </_.Container>
   );
 };
+
 export default ProfileEdit;
